+++ src/traits/insert/insert_interface.rs
@@ -3,7 +3,7 @@ use anyhow::Result;
 
 pub trait InsertInterface<T: Insertable>: Send + Sync
 {
-    fn insert(&self, item: &T, indexing_statement: Option<&str>) -> Result<u32>;
+    fn insert(&self, item: &T, indexing_statement: Option<&str>) -> Result<u64>;
 
     fn insert_and_fetch(&self, item: &T, indexing_statement: Option<&str>) -> Result<T>;
 }
