+++ src/traits/query/query_interface.rs
@@ -7,7 +7,7 @@ pub trait QueryInterface<T: Table>: Send + Sync
 
     fn query_all(&self) -> Result<Vec<T>>;
 
-    fn query_by_id(&self, id: u32) -> Result<T>;
+    fn query_by_id(&self, id: u64) -> Result<T>;
 
     fn query_by_expression(&self, expression: &str) -> Result<Vec<T>>;
 }
