+++ src/traits/update/update_interface.rs
@@ -3,5 +3,5 @@ use anyhow::Result;
 
 pub trait UpdateInterface<T: Updatable>: Send + Sync
 {
-    fn update_by_id(&self, id: u32, items: Vec<(String, String)>) -> Result<()>;
+    fn update_by_id(&self, id: u64, items: Vec<(String, String)>) -> Result<()>;
 }
